requirements.txt
psycopg[binary]
yfinance
python-dotenv